refactor: rename patternMatchers.{js,ts}
