+++ packages/patterns/src/type-from-pattern.ts
@@ -426,7 +426,7 @@ export type TypeFromInterfaceGuard<G> =
 // - matches: type predicate for narrowing in if-blocks
 // - mustMatch: asserts signature for narrowing after call
 //
-// The runtime implementations live in patternMatchers.js (which uses
+// The runtime implementations live in patternMatchers.ts (which uses
 // @ts-nocheck and cannot be directly type-checked) and are re-exported
 // by types-index.js.  These declarations overlay the runtime types so
 // that consumers get type-narrowing for free.
@@ -467,7 +467,7 @@ export namespace M {
 /**
  * Type-narrowing `matches`: narrows `specimen` in if-blocks.
  *
- * This declaration overlays the runtime `matches` from `patternMatchers.js`
+ * This declaration overlays the runtime `matches` from `patternMatchers.ts`
  * with a type-predicate signature.  The base runtime signature
  * `(specimen: unknown, patt: Pattern) => boolean` is verified as a supertype
  * in `types-index.js`.
