+++ packages/patterns/test/copySet.test.js
@@ -21,7 +21,7 @@ import {
   setIntersection,
   setDisjointSubtract,
 } from '../src/keys/merge-set-operators.js';
-import { M, matches } from '../src/patterns/patternMatchers.js';
+import { M, matches } from '../src/patterns/patternMatchers.ts';
 
 const { arbKey, exampleAlice, exampleBob, exampleCarol } = makeArbitraries(fc);
