+++ packages/patterns/test/mismatch-stack-demo.test-verbose.js
@@ -1,7 +1,7 @@
 import test from '@endo/ses-ava/test.js';
 
 import harden from '@endo/harden';
-import { M, mustMatch } from '../src/patterns/patternMatchers.js';
+import { M, mustMatch } from '../src/patterns/patternMatchers.ts';
 
 test('mismatch to see stack', t => {
   const err = t.throws(
