+++ packages/patterns/test/patterns.test.js
@@ -12,7 +12,7 @@ import {
   isKey,
   assertKey,
 } from '../src/keys/checkKey.js';
-import { mustMatch, matches, M } from '../src/patterns/patternMatchers.js';
+import { mustMatch, matches, M } from '../src/patterns/patternMatchers.ts';
 
 const { stringify: q } = JSON;
